WIP: clean up (4) & move helper functions into respective folders
