+++ scripts/shared/helper-functions.js
@@ -1,0 +1,21 @@
+import fs from 'fs';
+import ncp from 'ncp';
+import { promisify } from 'util';
+
+const copy = promisify(ncp);
+
+export const copyTemplateFiles = async (options) => {
+  return copy(options.templateDirectory, options.targetDirectory, {
+    clobber: false
+  });
+}
+
+export const deletePreviousTemplateFiles = async (filesArray, folderPath) => {
+  try {
+    filesArray.map((file) => {
+      fs.existsSync(`${folderPath}/${file}`) ? fs.unlinkSync(`${folderPath}/${file}`) : null;
+    });
+  } catch(err) {
+    console.log(err);
+  }
+}
